split components, refactored the code & deleted unnecessary console.logs
