+++ components/AddedExercisesList/AddedExercise/index.js
@@ -1,17 +1,7 @@
 import OpenCard from "./openCard";
 import { useState } from "react";
 
-export default function AddedExercisesList({ addedExercise }) {
-  return (
-    <ul>
-      {addedExercise.map((exercise) => (
-        <AddedExercise exercise={exercise} key={exercise.id} />
-      ))}
-    </ul>
-  );
-}
-
-function AddedExercise({ exercise }) {
+export default function AddedExercise({ exercise }) {
   const [openCard, setOpenCard] = useState(false);
 
   function openCardSetter() {
