+++ pages/index.js
@@ -1,7 +1,6 @@
 import ExerciseResults from "@/components/exercise-results";
 import ExerciseSearchForm from "@/components/exercise-search-form";
-import AddedExercises from "@/components/added-exercises";
-import AddedExercisesList from "@/components/added-exercises";
+import AddedExercisesList from "@/components/AddedExercisesList";
 
 export default function HomePage({
   setExerciseInput,
